ccp-provisioner/requirements.txt
requests==2.21.0
paho-mqtt==1.4.0